pyproject.toml
[project]
name = "hotel-booking"
version = "0.1.0"
description = ""
authors = [
    {name = "twentytwo-sixteen",email = "gadiev.mingiyan@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "hotel_booking", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
pre-commit = "^4.2.0"
alembic = "^1.16.1"
black = "^25.1.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
